requirements.txt
watchtower
aiodns==2.0.0
aiohttp==3.7.3
alabaster==0.7.12
amqp==5.0.1
appdirs==1.4.4
appnope==0.1.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
Authlib==0.15.3
autopep8==1.5.6
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
black==21.5b1
blessed==1.18.0
boto3==1.17.55
botocore==1.20.55
cachetools==4.2.1
celery==5.0.0
celery-progress==0.0.14
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Collectfast==2.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
decorator==4.4.2
defusedxml==0.6.0
dj-stripe==2.4.1
Django==3.1.6
django-allauth==0.44.0
django-anymail==8.1
django-braces==1.14.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-crontab==0.7.1
django-debug-toolbar==3.2.1
django-debug-toolbar-request-history==0.1.3
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-google-api==1.0.1
django-ipware==1.1.6
django-livereload-server==0.3.2
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-storages==1.11.1
django-timezone-field==4.1.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
docutils==0.16
drf-yasg==1.20.0
flyingcircus==0.1.3.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
idna-ssl==1.1.0
imagesize==1.2.0
imap-tools==0.40.0
IMAPClient==2.2.0
importlib-metadata==2.0.0
inflection==0.5.1
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonfield==3.1.0
jwcrypto==0.8
kombu==5.0.2
lxml==4.6.3
MarkupSafe==1.1.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
progressbar2==3.53.1
prompt-toolkit==3.0.8
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybuilder==0.12.10
pycares==3.1.1
pycodestyle==2.7.0
pycparser==2.20
pycurl==7.43.0.6
pydot==1.4.2
Pygments==2.7.4
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-utils==2.5.6
python3-openid==3.2.0
pytk==0.0.2.1
pytracking==0.2.1
pytz==2021.1
pyzmail36==1.0.4
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.1
sentry-sdk==1.0.0
setuptools-scm==6.0.1
six==1.15.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.54.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing==3.7.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zipp==3.3.0
dnspython==2.1.0
ipwhois==1.2.0